Add Dropdown of designation  On Root page

+++ app/javascript/packs/index.js
@@ -1,5 +1,6 @@
 import $ from 'jquery'
 import Rails from "@rails/ujs"
+
 $(function(){
   $('#delete_button').click(function(){
     var index_array = [];
@@ -10,14 +11,33 @@ $(function(){
       alert('plz select  some record !')
     }else{
       console.log(index_array)
-      $.ajax({
-        url: "/employees/employeesDestroy",
-        type: "DELETE", 
-        dataType: "script",
-        data: {"index_array": index_array}
-        // success: function(repsonse){...},
-        // error: function(repsonse){...}
-      })
+      if(confirm('You are removing employees record ?')){
+        $.ajax({
+          url: "/employees/employeesDestroy",
+          type: "DELETE", 
+          dataType: "script",
+          data: {"index_array": index_array}
+          // success: function(repsonse){...},
+          // error: function(repsonse){...}
+        });
+      }
     }
   });
 });
+
+
+$(function(){
+  $('#designations').change(function(){
+    var f = $("#designations").val();
+    var token = $('meta[name=csrf-token]').attr('first_name')
+    $.ajax({
+          url: "/employees/",
+          type: "GET", 
+          dataType: "script",
+          beforeSend: function(xhr) {xhr.setRequestHeader('X-CSRF-Token',token)},
+          data: {"designate": f}
+          // success: function(repsonse){...},
+          // error: function(repsonse){...}           
+    });
+  })
+})
